datasets/requirements.txt
python-dotenv
datasets
huggingface_hub
